verify_pypi/verify_pypi/apple_appstore_connect_client-1.0.4/requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-timeout>=2.1.0
python-dotenv>=1.0.0

# Linting and formatting
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
types-requests>=2.31.0
pandas-stubs>=2.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Packaging
build>=0.10.0
twine>=4.0.0